188_click_debounce_moved_idea_deleted
