+++ .gitignore
@@ -8,4 +8,4 @@
 .cxx
 local.properties
 build/
-.idea/
+.idea
