+++ app/src/main/java/com/practicum/playlistmaker/presentation/ui/search/TrackAdapter.kt
@@ -8,17 +8,13 @@ import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import com.practicum.playlistmaker.R
 import com.practicum.playlistmaker.domain.searchTracks.models.Track
+import com.practicum.playlistmaker.utils.constants.Constants
 
 class TrackAdapter(
     private val tracks: List<Track>
 ) : RecyclerView.Adapter<TrackViewHolder>() {
 
-    companion object {
-        private const val CLICK_DEBOUNCE_DELAY = 1000L
-    }
-
     var onItemClickListener: ((Track) -> Unit)? = null
-    private var isClickAllowed = true
 
     private val handler = Handler(Looper.getMainLooper())
 
@@ -31,23 +27,12 @@ class TrackAdapter(
         holder.bind(tracks[position])
         //слушатель нажатия на трек в результатах поиска
         holder.itemView.setOnClickListener {
-            if (clickDebounce()) {
-                //передача трека в активити
-                val track: Track = tracks[position]
-                onItemClickListener?.invoke(track)
-            }
+            //передача трека в активити
+            val track: Track = tracks[position]
+            onItemClickListener?.invoke(track)
         }
     }
 
     override fun getItemCount() = tracks.size
 
-
-    private fun clickDebounce() : Boolean {
-        val current = isClickAllowed
-        if (isClickAllowed) {
-            isClickAllowed = false
-            handler.postDelayed({isClickAllowed = true}, CLICK_DEBOUNCE_DELAY)
-        }
-        return current
-    }
 }
